ml/requirements.txt
kafka-python
scikit-learn==0.21.3
nltk==3.4.3